Split withFixedColumns Slate plugin into multiple files
REDMINE-20893
