+++ entry_types/scrolled/package/src/frontend/inlineEditing/EditableTable/withFixedColumns/index.js
@@ -1,0 +1,2 @@
+export {withFixedColumns} from './plugin';
+export {handleTableNavigation} from './handleTableNavigation';
