+++ entry_types/scrolled/package/src/frontend/inlineEditing/EditableTable/withFixedColumns/plugin.js
@@ -1,5 +1,8 @@
 import {Editor, Node, Path, Range, Transforms} from 'slate';
 
+import {Cell, CellPath, Row} from './helpers';
+import {TableTransforms, CellTransforms} from './transforms';
+
 export function withFixedColumns(editor) {
   const {deleteBackward, deleteForward, deleteFragment} = editor;
 
@@ -253,215 +256,4 @@ export function withFixedColumns(editor) {
   };
 
   return editor;
-}
-
-const CellTransforms = {
-  replaceContent(editor, nodes, {cellPath}) {
-    Transforms.insertText(editor, '', {at: cellPath});
-
-    if (nodes.length > 0) {
-      Transforms.insertNodes(editor, nodes, {at: [...cellPath, 0]});
-    }
-  },
-
-  deleteContentUntil(editor, {cellPath, point}) {
-    if (!Editor.isStart(editor, point, cellPath)) {
-      Transforms.delete(editor, {
-        at: {
-          anchor: Editor.start(editor, cellPath),
-          focus: point,
-        },
-      });
-    }
-  },
-
-  deleteContentFrom(editor, {cellPath, point}) {
-    if (!Editor.isEnd(editor, point, cellPath)) {
-      Transforms.delete(editor, {
-        at: {
-          anchor: point,
-          focus: Editor.end(editor, cellPath),
-        },
-      });
-    }
-  }
-};
-
-const TableTransforms = {
-  deleteRange(editor, [startPoint, endPoint]) {
-    const startCellMatch = Cell.match(editor, {at: startPoint.path});
-    const endCellMatch = Cell.match(editor, {at: endPoint.path});
-
-    if (startCellMatch && endCellMatch) {
-      const [startCellNode, startCellPath] = startCellMatch;
-      const [endCellNode, endCellPath] = endCellMatch;
-
-      if (!Path.equals(startCellPath, endCellPath)) {
-        const rewrittenCellPath = getRewrittenCellPath(startCellPath, endCellPath);
-
-        const rows = Array.from(Editor.nodes(editor, {
-          match: (n) => n.type === 'row',
-          at: { anchor: startPoint, focus: endPoint },
-        }));
-
-        CellTransforms.deleteContentFrom(editor, {
-          cellPath: startCellPath,
-          point: startPoint
-        });
-
-        if (rewrittenCellPath) {
-          const beforeNodes =
-            CellPath.columnIndex(startCellPath) === CellPath.columnIndex(endCellPath) ?
-            Cell.splitChildren(editor, {
-              cellNode: startCellNode,
-              point: startPoint
-            })[0] :
-            [];
-
-          const [, afterNodes] = Cell.splitChildren(editor, {
-            cellNode: endCellNode,
-            point: endPoint
-          });
-          CellTransforms.replaceContent(editor, beforeNodes.concat(afterNodes), {
-            cellPath: rewrittenCellPath
-          })
-          Transforms.select(editor, {
-            ...Editor.start(editor, rewrittenCellPath),
-            offset: beforeNodes[beforeNodes.length - 1]?.text.length || 0
-          });
-
-          rows.reverse().forEach(([_, rowPath]) => {
-            if (rowPath[rowPath.length - 1] !== CellPath.rowIndex(rewrittenCellPath)) {
-              Transforms.removeNodes(editor, {at: rowPath});
-            }
-          });
-        }
-        else {
-          CellTransforms.deleteContentUntil(editor, {
-            cellPath: endCellPath,
-            point: endPoint
-          });
-
-          rows.slice(1, -1).reverse().forEach(([_, rowPath]) => {
-            Transforms.removeNodes(editor, {at: rowPath});
-          });
-
-          Transforms.select(editor, startPoint);
-        }
-
-        return;
-      }
-    }
-
-    function getRewrittenCellPath(startCellPath, endCellPath) {
-      if (CellPath.columnIndex(startCellPath) < CellPath.columnIndex(endCellPath)) {
-        const [, rewrittenCellPath] = Editor.next(editor, {at: startCellPath});
-        return rewrittenCellPath;
-      }
-      else if (CellPath.columnIndex(startCellPath) > CellPath.columnIndex(endCellPath)) {
-        return null;
-      }
-      else if (CellPath.columnIndex(startCellPath) === 0) {
-        return endCellPath;
-      }
-      else {
-        return startCellPath;
-      }
-    }
-  }
-};
-
-const Row = {
-  match(editor) {
-    const [rowMatch] = Editor.nodes(editor, {
-      match: (n) => n.type === 'row',
-    });
-
-    return rowMatch;
-  }
-}
-
-const Cell = {
-  splitChildren(editor, {cellNode, point}) {
-    const [leafNode, leafPath] = Editor.leaf(editor, point.path);
-
-    const cursorOffset = point.offset;
-    const text = leafNode.text || '';
-    const beforeText = text.slice(0, cursorOffset);
-    const afterText = text.slice(cursorOffset);
-
-    const splitIndex = leafPath[leafPath.length - 1];
-
-    const beforeNodes = [];
-    const afterNodes = [];
-
-    cellNode.children.forEach((node, index) => {
-      if (index < splitIndex) {
-        beforeNodes.push(node);
-      } else if (index === splitIndex) {
-        if (beforeText) {
-          beforeNodes.push({ ...node, text: beforeText });
-        }
-        if (afterText) {
-          afterNodes.push({ ...node, text: afterText });
-        }
-      } else {
-        afterNodes.push(node);
-      }
-    });
-
-    return [beforeNodes, afterNodes];
-  },
-
-  match(editor, {at} = {}) {
-    const [cellMatch] = Editor.nodes(editor, {
-      match: n => n.type === 'label' || n.type === 'value',
-      at
-    });
-
-    return cellMatch;
-  }
-}
-
-const CellPath = {
-  columnIndex(path) {
-    return path[path.length - 1];
-  },
-
-  rowIndex(path) {
-    return path[path.length - 2];
-  }
-};
-
-export function handleTableNavigation(editor, event) {
-  const {selection} = editor;
-
-  if (selection && Range.isCollapsed(selection)) {
-    const cellMatch = Cell.match(editor);
-
-    if (cellMatch) {
-      const [, cellPath] = cellMatch;
-      const rowPath = cellPath.slice(0, -1);
-
-      if (event.key === 'ArrowUp') {
-        event.preventDefault();
-
-        if (rowPath[rowPath.length - 1] > 0) {
-          const previousRowPath = Path.previous(rowPath);
-          const targetPath = [...previousRowPath, cellPath[cellPath.length - 1]];
-
-          Transforms.select(editor, Editor.start(editor, targetPath));
-        }
-      } else if (event.key === 'ArrowDown') {
-        event.preventDefault();
-
-        const nextRowPath = Path.next(rowPath);
-        const targetPath = [...nextRowPath, cellPath[cellPath.length - 1]];
-
-        if (Node.has(editor, targetPath)) {
-          Transforms.select(editor, Editor.start(editor, targetPath));
-        }
-      }
-    }
-  }
 }
