+++ entry_types/scrolled/package/src/frontend/inlineEditing/EditableTable/withFixedColumns/transforms/index.js
@@ -1,0 +1,2 @@
+export {TableTransforms} from './table';
+export {CellTransforms} from './cell';
